+++ BluebirdPS/append/InitializeClient.ps1
@@ -24,10 +24,9 @@ Register-ArgumentCompleter -CommandName Get-TwitterApiEndpoint -ParameterName Co
 
 # store EndpointInfo in module variable
 $BluebirdPSCommands = Get-Command -Module BluebirdPS -ListImported
-$PublicFunctions = (Get-Module -Name BluebirdPS).ExportedFunctions.Values.Name
 
 [SuppressMessage('PSUseDeclaredVarsMoreThanAssigments', '')]
-$TwitterEndpoints = foreach ($Command in $BluebirdPSCommands) {
+$script:TwitterEndpoints = foreach ($Command in $BluebirdPSCommands) {
     $NavigationLinks = (Get-Help -Name $Command.Name).relatedLinks.navigationLink.Where{$_.linkText -match '^(?!.*(Online|\w+-)).*$'}.Where{$_.linkText -match '- \w+\s(\/|\w+\/)'}
     if ($NavigationLinks.Count -gt 0) {
         $ApiEndpoint = $NavigationLinks.LinkText | ForEach-Object { $_.Split('-')[1].Trim() }
@@ -37,7 +36,6 @@ $TwitterEndpoints = foreach ($Command in $BluebirdPSCommands) {
     }
     [EndpointInfo]::new(
         $Command.Name,
-        ($Command.Name -notin $PublicFunctions ? 'Private' : 'Public'),
         $ApiEndpoint,
         $ApiDocumentation
     )