+++ BluebirdPS/public/api_v1/direct_message/Publish-TwitterDM.ps1
@@ -1,37 +1,42 @@
 function Publish-TwitterDM {
-    [CmdletBinding(DefaultParameterSetName='DM')]
+    [CmdletBinding(DefaultParameterSetName='DMUserId')]
     param(
-        [Parameter(Mandatory)]
-        [ValidateLength(1,10000)]
         [string]$Message,
 
-        [Parameter(Mandatory,ParameterSetName='DM',ValueFromPipeline)]
-        [Parameter(Mandatory,ParameterSetName='DMWithMedia',ValueFromPipeline)]
+        [Parameter(Mandatory,ParameterSetName='DMUserId',ValueFromPipeline)]
+        [Parameter(Mandatory,ParameterSetName='DMUserIdWithMedia',ValueFromPipeline)]
         [string]$Id,
 
         [Parameter(Mandatory,ParameterSetName='DMUserObject',ValueFromPipeline)]
-        [Parameter(Mandatory,ParameterSetName='DMWithMedia',ValueFromPipeline)]
+        [Parameter(Mandatory,ParameterSetName='DMUserObjectWithMedia',ValueFromPipeline)]
+        [ValidateObjectNotNullOrEmpty()]
         [BluebirdPS.APIV2.UserInfo.User]$User,
 
-        [Parameter(ParameterSetName='DM')]
+        [Parameter(ParameterSetName='DMUserId')]
         [Parameter(ParameterSetName='DMUserObject')]
         [ValidateNotNullOrEmpty()]
         [string]$MediaId,
 
-        [Parameter(Mandatory,ParameterSetName='DMWithMedia')]
+        [Parameter(Mandatory,ParameterSetName='DMUserIdWithMedia')]
+        [Parameter(Mandatory,ParameterSetName='DMUserObjectWithMedia')]
         [ValidateScript({Test-Path -Path $_})]
         [string]$Path,
 
-        [Parameter(Mandatory,ParameterSetName='DMWithMedia')]
+        [Parameter(Mandatory,ParameterSetName='DMUserIdWithMedia')]
+        [Parameter(Mandatory,ParameterSetName='DMUserObjectWithMedia')]
         [ValidateSet('DMImage','DMVideo','DMGif')]
         [string]$Category,
 
-        [Parameter(ParameterSetName='DMWithMedia')]
+        [Parameter(ParameterSetName='DMUserIdWithMedia')]
+        [Parameter(ParameterSetName='DMUserObjectWithMedia')]
         [ValidateLength(1,1000)]
         [string]$AltImageText
     )
 
-    if ($PSCmdlet.ParameterSetName -eq 'DMWithMedia') {
+    $MessageTemplate = '{{"event":{{"type":"message_create","message_create":{{"target":{{"recipient_id":{0}}},"message_data":{{"text":"{1}"}}}}}}}}'
+    $MessageWithMediaTemplate = '{{"event":{{"type":"message_create","message_create":{{"target":{{"recipient_id":{0}}},"message_data":{{"text":"{1}","attachment":{{"type":"media","media":{{"id":{2}}}}}}}}}}}}}'
+
+    if ($PSCmdlet.ParameterSetName -match 'WithMedia') {
         $TwitterMediaParams = @{
             Path = $Path
             Category = $Category
@@ -42,13 +47,18 @@ function Publish-TwitterDM {
         $MediaId = Send-TwitterMedia @TwitterMediaParams | Select-Object -ExpandProperty media_id
     }
 
-    $MessageTemplate = '{{"event":{{"type":"message_create","message_create":{{"target":{{"recipient_id":"{0}"}},"message_data":{{"text":"{1}"}}}}}}}}'
-    $MessageWithMediaTemplate = '{{"event":{{"type":"message_create","message_create":{{"target":{{"recipient_id":"{0}"}},"message_data":{{"text":"{1}","attachment":{{"type":"media","media":{{"id":{2}}}}}}}}}}}}}'
+    $RecipientId = $PSCmdlet.ParameterSetName -match 'DMUserObject' ? $User.Id : $Id
+    $MessageText = [string]::IsNullOrEmpty($Message) ? [string]::Empty : $Message
 
-    if ($MediaId) {
-        $Body = $MessageWithMediaTemplate -f $Id,$Message,$MediaId
+    if ($MessageText) {
+        if ($MediaId) {
+            $Body = $MessageWithMediaTemplate -f $RecipientId,$MessageText,$MediaId
+        } else {
+            $Body = $MessageTemplate -f $RecipientId,$MessageText
+        }
     } else {
-        $Body = $MessageTemplate -f $Id,$Message
+        'You must provide a message, media, or a message and media. Please try again.' | Write-Warning
+        return
     }
 
     $Request = [TwitterRequest]@{