+++ BluebirdPS/public/api_v1/lists/Get-TwitterListSubscriber.ps1
@@ -6,6 +6,7 @@ function Get-TwitterListSubscriber {
         [string]$Id,
 
         [Parameter(Mandatory,ParameterSetName='ByList',ValueFromPipeline)]
+        [ValidateObjectNotNullOrEmpty()]
         [BluebirdPS.APIV1.List]$List
     )