+++ BluebirdPS/public/api_v1/lists/Remove-TwitterListSubscription.ps1
@@ -6,6 +6,7 @@ function Remove-TwitterListSubscription {
         [string]$Id,
 
         [Parameter(Mandatory,ParameterSetName='ByList',ValueFromPipeline)]
+        [ValidateObjectNotNullOrEmpty()]
         [BluebirdPS.APIV1.List]$List
     )