+++ BluebirdPS/public/api_v1/lists/Set-TwitterList.ps1
@@ -6,6 +6,7 @@ function Set-TwitterList {
         [string]$Id,
 
         [Parameter(Mandatory,ParameterSetName='ByList',ValueFromPipeline)]
+        [ValidateObjectNotNullOrEmpty()]
         [BluebirdPS.APIV1.List]$List,
 
         [Parameter()]