+++ BluebirdPS/public/api_v1/lists/Test-TwitterListSubscription.ps1
@@ -6,6 +6,7 @@ function Test-TwitterListSubscription {
         [string]$Id,
 
         [Parameter(Mandatory,ParameterSetName='ByList',ValueFromPipeline)]
+        [ValidateObjectNotNullOrEmpty()]
         [BluebirdPS.APIV1.List]$List,
 
         [ValidateNotNullOrEmpty()]