+++ BluebirdPS/public/api_v1/tweets/Unpublish-Tweet.ps1
@@ -4,6 +4,7 @@ function Unpublish-Tweet {
         [Parameter(Mandatory,ParameterSetName='ById',ValueFromPipeline)]
         [string]$Id,
         [Parameter(Mandatory,ParameterSetName='ByTweet',ValueFromPipeline)]
+        [ValidateObjectNotNullOrEmpty()]
         [BluebirdPS.APIV2.TweetInfo.Tweet]$Tweet
     )