+++ BluebirdPS/public/api_v2/tweets/Get-TweetPoll.ps1
@@ -8,6 +8,7 @@ function Get-TweetPoll {
         [string]$Id,
 
         [Parameter(Mandatory,ValueFromPipeline,ParameterSetName='ByTweet')]
+        [ValidateObjectNotNullOrEmpty()]
         [BluebirdPS.APIV2.TweetInfo.Tweet]$Tweet
     )