+++ BluebirdPS/public/api_v2/users_followers_friends/Get-TwitterFollowers.ps1
@@ -6,6 +6,7 @@ function Get-TwitterFollowers {
         [string]$Id,
 
         [Parameter(Mandatory,ParameterSetName='ByUser',ValueFromPipeline)]
+        [ValidateObjectNotNullOrEmpty()]
         [BluebirdPS.APIV2.UserInfo.User]$User,
 
         [switch]$IncludeExpansions