+++ BluebirdPS/public/api_v2/users_followers_friends/Remove-TwitterFriend.ps1
@@ -5,6 +5,7 @@ function Remove-TwitterFriend {
         [string]$Id,
 
         [Parameter(Mandatory,ParameterSetName='ByUser',ValueFromPipeline)]
+        [ValidateObjectNotNullOrEmpty()]
         [BluebirdPS.APIV2.UserInfo.User]$User
     )