+++ BluebirdPS/public/helpers/Set-BluebirdPSConfiguration.ps1
@@ -3,7 +3,7 @@ function Set-BluebirdPSConfiguration {
     param(
         [BluebirdPS.RateLimitAction]$RateLimitAction,
         [int]$RateLimitThreshold,
-        [bool]$RawOutput,
+        [BluebirdPS.OutputType]$OutputType,
         [switch]$Export
     )