+++ src/APIV2/TweetInfo.cs
@@ -2,6 +2,7 @@
 using BluebirdPS.APIV2.TweetInfo.Metrics;
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Management.Automation;
 
 namespace BluebirdPS.APIV2.TweetInfo
@@ -118,6 +119,10 @@ namespace BluebirdPS.APIV2.TweetInfo
             }
         }
 
+        public override string ToString()
+        {
+            return $"{Id}:{Text}";
+        }
     }
 
     public class ReferencedTweet : TwitterObject
@@ -131,6 +136,51 @@ namespace BluebirdPS.APIV2.TweetInfo
             Type = Helpers.ToTitleCase(input.type);
             Id = input.id;
             OriginalObject = input;
+        }
+    }
+
+    public class TweetCount : TwitterObject
+    {
+        public DateTime Start { get; set; }
+        public DateTime End { get; set; }
+        public long Count { get; set; }
+
+        public TweetCount(dynamic input)
+        {
+            OriginalObject = input;
+
+            Start = input.start.ToLocalTime();
+            End = input.end.ToLocalTime();
+            Count = input.tweet_count;
+        }
+
+        public override string ToString()
+        {
+            return $"Start: {Start.ToString("G",DateTimeFormatInfo.InvariantInfo)}, End: {End.ToString("G", DateTimeFormatInfo.InvariantInfo)}, Count: {Count}";
+        }
+    }
+
+    public class TweetCountSummary
+    {
+        public string SearchString { get; set; }
+        public string Granularity { get; set; }
+        public DateTime StartTime { get; set; }
+        public DateTime EndTime { get; set; }        
+        public long TotalCount { get; set; }
+
+        public TweetCountSummary() { }
+        public TweetCountSummary(string search, string granularity, DateTime starttime, DateTime endtime, long totalCount)
+        {
+            SearchString = search;
+            Granularity = granularity;
+            StartTime = starttime.ToLocalTime();
+            EndTime = endtime.ToLocalTime();            
+            TotalCount = totalCount;
+        }
+
+        public override string ToString()
+        {
+            return $"SearchString: {SearchString}, Granularity: {Granularity}, TotalCount: {TotalCount}, StartTime: {StartTime.ToString("G", DateTimeFormatInfo.InvariantInfo)}, EndTime: {EndTime.ToString("G", DateTimeFormatInfo.InvariantInfo)}";
         }
     }